Move process-monitor into its own file
